Share common files between cannonkey boards
