--- keyboards/cannonkeys/ortho48/led.c
@@ -1,51 +1,0 @@
-/*
-Copyright 2012 Jun Wako <wakojun@gmail.com>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include "hal.h"
-#include "backlight.h"
-#include "led.h"
-#include "printf.h"
-
-void backlight_init_ports(void) {
-    printf("backlight_init_ports()\n");
-    #ifdef BACKLIGHT_ENABLE
-    palSetPadMode(GPIOA, 8, PAL_MODE_OUTPUT_PUSHPULL);
-    palSetPad(GPIOA, 8);
-    #endif
-}
-
-void backlight_set(uint8_t level) {
-    printf("backlight_set(%d)\n", level);
-    #ifdef BACKLIGHT_ENABLE
-    if (level == 0) {
-        // Turn backlight off
-        palSetPad(GPIOA, 8);
-    } else {
-        // Turn backlight on
-        palClearPad(GPIOA, 8);
-    }
-    #endif
-}
-
-void led_set(uint8_t usb_led)
-{
-    if (usb_led & (1<<USB_LED_CAPS_LOCK)) {
-        palClearPad(GPIOC, 13);
-    } else {
-    	palSetPad(GPIOC, 13);
-    }
-}
