+++ keyboards/cannonkeys/ortho60/readme.md
@@ -7,6 +7,6 @@ Hardware Supported: Blue Pill STM32F103C8T6
 
 Make example for this keyboard (after setting up your build environment):
 
-    make handwired/ortho60:default
+    make cannonkeys/ortho60:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
