+++ keyboards/cannonkeys/ortho60/rules.mk
@@ -1,7 +1,8 @@
 # project specific files
+VPATH += keyboards/cannonkeys/bluepill
 SRC =	led.c \
-		  underglow.c \
-		  hsv2rgb.c
+		  ws2812.c \
+      keyboard.c
 
 # GENERIC STM32F103C8T6 board - stm32duino bootloader
 OPT_DEFS = -DCORTEX_VTOR_INIT=0x2000
@@ -50,5 +51,6 @@ COMMAND_ENABLE = yes    # Commands for debug and configuration
 SLEEP_LED_ENABLE = yes  # Breathing sleep LED during USB suspend
 NKRO_ENABLE = yes	    # USB Nkey Rollover
 BACKLIGHT_ENABLE = yes
+RGBLIGHT_ENABLE = yes
 
 LAYOUTS = ortho_5x12
