+++ keyboards/cannonkeys/practice60/practice60.c
@@ -1,35 +1,2 @@
 
 #include "practice60.h"
-
-#include "ch.h"
-#include "hal.h"
-#include "led_custom.h"
-#include "print.h"
-#include "debug.h"
-#include "util.h"
-
-/* generic STM32F103C8T6 board */
-#ifdef BOARD_GENERIC_STM32_F103
-#define LED_ON()    do { palClearPad(GPIOC, 13) ;} while (0)
-#define LED_OFF()   do { palSetPad(GPIOC, 13); } while (0)
-#define LED_TGL()   do { palTogglePad(GPIOC, 13); } while (0)
-#endif
-
-void matrix_init_kb(void){
-      /* MOSI pin*/
-    palSetPadMode(GPIOB, 15, PAL_MODE_STM32_ALTERNATE_PUSHPULL);
-    LED_ON();
-    wait_ms(500);
-    LED_OFF();
-
-#ifdef RGBLIGHT_ENABLE
-    leds_init();
-#endif
-}
-
-void matrix_scan_kb(void)
-{
-  #ifdef RGBLIGHT_ENABLE
-    rgblight_task();
-  #endif
-}
