+++ keyboards/cannonkeys/practice60/readme.md
@@ -1,3 +1,12 @@
 # Practice 60
 
-Blue Pill STM32F103C8T6 based 60% board
+A Blue Pill STM32F103C8T6-based 60% ANSI board.
+
+Keyboard Maintainer: [Andrew Kannan](https://github.com/awkannan1)  
+Hardware Supported: Blue Pill STM32F103C8T6  
+
+Make example for this keyboard (after setting up your build environment):
+
+    make cannonkeys/practice60:default
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
